Build27
  - Improved theme change animation in theme selection panel

+++ app/src/main/java/app/simple/inure/themes/interfaces/ThemeRevealCoordinatesListener.java
@@ -1,5 +1,5 @@
 package app.simple.inure.themes.interfaces;
 
 public interface ThemeRevealCoordinatesListener {
-    void onTouchCoordinates(float x, float y);
+    void onTouchCoordinates(int x, int y);
 }
